+++ packages/react-devtools/src/components/OverviewTab.tsx
@@ -17,7 +17,9 @@
 import { AnchorButton, NonIdealState } from "@blueprintjs/core";
 import React from "react";
 
-import { useOverviewMetrics } from "../hooks/useOverviewMetrics.js";
+import { useClientMetrics } from "../hooks/useClientMetrics.js";
+import { useComponentOntology } from "../hooks/useComponentOntology.js";
+import { useDebuggingTiles } from "../hooks/useDebuggingTiles.js";
 import { formatMetric } from "../metrics/clientMetrics.js";
 import type { MonitorStore } from "../store/MonitorStore.js";
 import { Metric } from "./Metric.js";
@@ -172,4 +174,23 @@ export function OverviewTab({
       </OverviewSection>
     </div>
   );
+}
+
+function useOverviewMetrics(monitorStore: MonitorStore) {
+  const clientMetrics = useClientMetrics(monitorStore);
+  const { overfetchingCount, errorWarningCount } =
+    useDebuggingTiles(monitorStore);
+  const { facets } = useComponentOntology(monitorStore);
+
+  return {
+    cacheHitRate: clientMetrics.cacheHitRate,
+    requestsSaved: clientMetrics.requestsSaved,
+    estimatedTimeSavedMs: clientMetrics.estimatedTimeSavedMs,
+    avgResponseMs: clientMetrics.avgResponseMs,
+    objectTypeCount: facets.objectTypes.length,
+    actionTypeCount: facets.actions.length,
+    linkCount: facets.links.length,
+    errorWarningCount,
+    overfetchingCount,
+  };
 }
